Moved toTakerFunction to takers
