+++ src/main/rules/createRule.ts
@@ -1,6 +1,5 @@
-import {Taker} from '../takers';
+import {Taker, toTakerFunction} from '../takers';
 import {Rule, StageProvider} from './rule-types';
-import {toTakerFunction} from './rule-utils';
 
 /**
  * Creates the new rule that reads tokens.
