+++ src/main/rules/index.ts
@@ -1,4 +1,3 @@
 export * from './compileRuleIterator';
 export * from './createRule';
 export * from './rule-types';
-export {toTakerFunction} from './rule-utils';
